refactor(cmd): extract dispatch and resolution (#525)
