+++ lua/forge/cmd.lua
@@ -2,7 +2,6 @@ local M = {}
 local collections = require('forge.collections')
 
 local detect = require('forge.detect')
-local ops = require('forge.ops')
 local state_mod = require('forge.state')
 local surface = require('forge.surface')
 
@@ -309,12 +308,6 @@ local function target_parse_opts()
   return require('forge.target').parse_opts()
 end
 
----@param value forge.TargetValue|forge.HeadInput|forge.Scope|nil
----@return forge.RepoLike?
-local function repo_target(value)
-  return require('forge.target').repo_target(value)
-end
-
 local function default_policy(command, _)
   local policy = {}
   if not command.parsed_modifiers.repo and command.family == 'pr' and command.name == 'create' then
@@ -358,428 +351,6 @@ local function default_targets(command, parse_opts)
   end
   return policy, defaults
 end
-
-local function require_git_or_warn()
-  vim.fn.system('git rev-parse --show-toplevel')
-  if vim.v.shell_error ~= 0 then
-    warn('not a git repository')
-    return false
-  end
-  return true
-end
-
-local function require_forge_or_warn()
-  local forge_mod = require('forge')
-  local f = forge_mod.detect()
-  if not f then
-    warn('no forge detected')
-    return nil, forge_mod
-  end
-  return f, forge_mod
-end
-
----@param command forge.Command
----@param forge_name forge.ScopeKind
----@return forge.Scope?
-local function resolve_scope_modifier(command, forge_name)
-  local target = require('forge.target')
-  local repo = repo_target(command.parsed_modifiers.target)
-    or repo_target(command.parsed_modifiers.base)
-    or repo_target(command.parsed_modifiers.head)
-    or repo_target(command.parsed_modifiers.repo)
-    or repo_target(command.default_targets.target)
-    or repo_target(command.default_targets.rev)
-    or repo_target(command.default_targets.base)
-    or repo_target(command.default_targets.head)
-    or repo_target(command.default_targets.repo)
-  return target.resolve_scope(repo, forge_name, target_parse_opts())
-end
-
-local function resolve_repo_modifier(command, forge_name)
-  local target = require('forge.target')
-  local repo = repo_target(command.parsed_modifiers.repo)
-    or repo_target(command.default_targets.repo)
-  return target.resolve_scope(repo, forge_name, target_parse_opts())
-end
-
-local function current_pr_resolution_opts(command, f)
-  local opts = {
-    forge = f,
-  }
-  if command.parsed_modifiers.repo ~= nil then
-    opts.repo = command.parsed_modifiers.repo
-  end
-  if command.parsed_modifiers.head ~= nil then
-    opts.head = command.parsed_modifiers.head
-  end
-  return opts
-end
-
-local function resolve_current_pr_or_warn(command, f)
-  local pr, err = require('forge').current_pr(current_pr_resolution_opts(command, f))
-  if err then
-    warn(err.message)
-    return nil
-  end
-  if pr then
-    return pr
-  end
-  warn(('no open %s found for this branch'):format((f.labels and f.labels.pr_one) or 'PR'))
-  return nil
-end
-
-local function resolve_branch_pr_or_warn(command, f, policy, no_match)
-  local pr, err = require('forge.resolve').branch_pr(current_pr_resolution_opts(command, f), policy)
-  if err then
-    warn(err.message)
-    return nil
-  end
-  if pr then
-    return pr
-  end
-  warn(no_match)
-  return nil
-end
-
-local function dispatch_pr(command)
-  if not require_git_or_warn() then
-    return
-  end
-  local f = require_forge_or_warn()
-  if not f then
-    return
-  end
-  local num = command.subjects[1]
-  if command.name == 'create' then
-    local target = require('forge.target')
-    local forge_mod = require('forge')
-    local parse_opts = target_parse_opts()
-    local head = command.parsed_modifiers.head or command.default_targets.head
-    local base = command.parsed_modifiers.base or command.default_targets.base
-    local scope = resolve_repo_modifier(command, f.name)
-    forge_mod.create_pr({
-      draft = command.modifiers.draft == true,
-      instant = command.modifiers.fill == true,
-      web = command.modifiers.web == true,
-      scope = scope,
-      head_branch = head and head.rev or nil,
-      head_scope = target.resolve_scope(head, f.name, parse_opts),
-      base_branch = base and base.rev or nil,
-      base_scope = target.resolve_scope(base, f.name, parse_opts) or scope,
-    })
-    return
-  end
-  if command.name == 'open' then
-    if num then
-      ops.pr_edit({ num = num, scope = resolve_repo_modifier(command, f.name) })
-      return
-    end
-    local pr = resolve_current_pr_or_warn(command, f)
-    if not pr then
-      return
-    end
-    ops.pr_edit(pr)
-    return
-  end
-  if command.name == 'ci' then
-    local pr = num and { num = num, scope = resolve_repo_modifier(command, f.name) }
-      or resolve_branch_pr_or_warn(command, f, {
-        searches = {
-          { 'open' },
-          { 'closed', 'merged' },
-        },
-      }, ('no %s found for this branch'):format((f.labels and f.labels.pr_one) or 'PR'))
-    if not pr then
-      return
-    end
-    ops.pr_ci(f, pr)
-    return
-  end
-  if command.name == 'browse' then
-    local scope = resolve_repo_modifier(command, f.name)
-    if num then
-      ops.pr_browse(f, { num = num, scope = scope })
-    else
-      ops.list_browse(f, 'pr', { scope = scope })
-    end
-    return
-  end
-  if command.name == 'refresh' then
-    state_mod.clear_list_kind('pr')
-    require('forge.logger').info('refreshed ' .. ((f.labels and f.labels.pr) or 'pr') .. ' list')
-    return
-  end
-  local function action_pr()
-    if num then
-      return { num = num, scope = resolve_repo_modifier(command, f.name) }
-    end
-    return resolve_current_pr_or_warn(command, f)
-  end
-  if command.name == 'edit' then
-    ops.pr_edit({ num = num, scope = resolve_repo_modifier(command, f.name) })
-    return
-  end
-  if command.name == 'approve' then
-    local pr = action_pr()
-    if not pr then
-      return
-    end
-    ops.pr_approve(f, pr)
-    return
-  end
-  if command.name == 'merge' then
-    local pr = action_pr()
-    if not pr then
-      return
-    end
-    ops.pr_merge(f, pr, command.modifiers.method)
-    return
-  end
-  if command.name == 'draft' then
-    local pr = action_pr()
-    if not pr then
-      return
-    end
-    ops.pr_toggle_draft(f, pr, false)
-    return
-  end
-  if command.name == 'ready' then
-    local pr = action_pr()
-    if not pr then
-      return
-    end
-    ops.pr_toggle_draft(f, pr, true)
-    return
-  end
-  if command.name == 'close' then
-    local pr = action_pr()
-    if not pr then
-      return
-    end
-    ops.pr_close(f, pr)
-    return
-  end
-  if command.name == 'reopen' then
-    local pr = num and { num = num, scope = resolve_repo_modifier(command, f.name) }
-      or resolve_branch_pr_or_warn(command, f, {
-        searches = {
-          { 'closed' },
-        },
-      }, ('no reopenable %s found for this branch'):format(
-        (f.labels and f.labels.pr_one) or 'PR'
-      ))
-    if not pr then
-      return
-    end
-    ops.pr_reopen(f, pr)
-    return
-  end
-  warn(('unsupported pr action: %s'):format(command.name))
-end
-
-local function dispatch_review(command)
-  if not require_git_or_warn() then
-    return
-  end
-  local f = require_forge_or_warn()
-  if not f then
-    return
-  end
-  local num = command.subjects[1]
-  local opts = {
-    adapter = command.modifiers.adapter ~= true and command.modifiers.adapter or nil,
-  }
-  if num then
-    ops.pr_review(f, { num = num, scope = resolve_repo_modifier(command, f.name) }, opts)
-    return
-  end
-  local pr = resolve_current_pr_or_warn(command, f)
-  if not pr then
-    return
-  end
-  ops.pr_review(f, pr, opts)
-end
-
-local function dispatch_issue(command)
-  if not require_git_or_warn() then
-    return
-  end
-  local f = require_forge_or_warn()
-  if not f then
-    return
-  end
-  local num = command.subjects[1]
-  local scope = resolve_repo_modifier(command, f.name)
-  if command.name == 'create' then
-    local template = command.modifiers.template
-    require('forge').create_issue({
-      web = command.modifiers.web == true,
-      blank = command.modifiers.blank == true,
-      template = template ~= true and template or nil,
-      scope = scope,
-    })
-    return
-  end
-  if command.name == 'browse' then
-    if num then
-      ops.issue_browse(f, { num = num, scope = scope })
-    else
-      ops.list_browse(f, 'issue', { scope = scope })
-    end
-    return
-  end
-  if command.name == 'edit' then
-    ops.issue_edit({ num = num, scope = scope })
-    return
-  end
-  if command.name == 'close' then
-    ops.issue_close(f, { num = num, scope = scope })
-    return
-  end
-  if command.name == 'reopen' then
-    ops.issue_reopen(f, { num = num, scope = scope })
-    return
-  end
-  if command.name == 'refresh' then
-    state_mod.clear_list_kind('issue')
-    require('forge.logger').info('refreshed issue list')
-    return
-  end
-  warn(('unsupported issue action: %s'):format(command.name))
-end
-
-local function dispatch_ci(command)
-  if not require_git_or_warn() then
-    return
-  end
-  if command.name == 'open' and command.subjects[1] == nil then
-    require('forge').ci({
-      repo = command.parsed_modifiers.repo,
-      head = command.parsed_modifiers.head,
-    })
-    return
-  end
-  local f = require_forge_or_warn()
-  if not f then
-    return
-  end
-  local scope = resolve_repo_modifier(command, f.name)
-  if command.name == 'open' then
-    ops.ci_open(f, { id = command.subjects[1], scope = scope })
-    return
-  end
-  if command.name == 'browse' then
-    local id = command.subjects[1]
-    if id then
-      ops.ci_browse(f, { id = id, scope = scope })
-    else
-      ops.list_browse(f, 'ci', { scope = scope })
-    end
-    return
-  end
-  if command.name == 'refresh' then
-    state_mod.clear_list_kind('ci')
-    require('forge.logger').info('refreshed CI run list')
-    return
-  end
-  warn(('unsupported ci action: %s'):format(command.name))
-end
-
-local function dispatch_release(command)
-  if not require_git_or_warn() then
-    return
-  end
-  local f = require_forge_or_warn()
-  if not f then
-    return
-  end
-  local tag = command.subjects[1]
-  local scope = resolve_repo_modifier(command, f.name)
-  if command.name == 'browse' then
-    if tag then
-      ops.release_browse(f, { tag = tag, scope = scope })
-    else
-      ops.list_browse(f, 'release', { scope = scope })
-    end
-    return
-  end
-  if command.name == 'delete' then
-    ops.release_delete(f, { tag = tag, scope = scope })
-    return
-  end
-  if command.name == 'refresh' then
-    state_mod.clear_list_kind('release')
-    require('forge.logger').info('refreshed release list')
-    return
-  end
-end
-
-local function dispatch_browse(command)
-  if not require_git_or_warn() then
-    return
-  end
-  local f = require_forge_or_warn()
-  if not f then
-    return
-  end
-  local scope = resolve_scope_modifier(command, f.name)
-  local subject = command.subjects[1]
-  local forge_mod = require('forge')
-  if subject then
-    ops.browse_subject(f, { num = subject, scope = scope })
-    return
-  end
-  local location = command.parsed_modifiers.target
-  if location then
-    local branch = location.rev and location.rev.rev or nil
-    if not branch then
-      local explicit_branch = command.parsed_modifiers.branch
-      branch = explicit_branch and explicit_branch.branch or nil
-    end
-    if not branch then
-      local ctx = forge_mod.current_context()
-      branch = type(ctx) == 'table' and ctx.branch or nil
-    end
-    if ops.browse_location(f, location, scope, branch) then
-      return
-    end
-    require('forge.logger').warn('detached HEAD')
-    return
-  end
-  local commit = command.parsed_modifiers.commit
-  if commit and commit.commit then
-    forge_mod.open('browse.commit', { commit = commit.commit, scope = scope })
-    return
-  end
-  local branch = command.parsed_modifiers.branch
-  local file_loc = forge_mod.file_loc(command.range)
-  if branch and branch.branch then
-    if ops.browse_file(f, file_loc, branch.branch, scope) then
-      return
-    end
-    forge_mod.open('browse.branch', { branch = branch.branch, scope = scope })
-    return
-  end
-  local ctx = forge_mod.current_context()
-  local ctx_branch = type(ctx) == 'table' and ctx.branch or nil
-  if ops.browse_file(f, file_loc, ctx_branch, scope) then
-    return
-  end
-  ops.browse_repo({ scope = scope })
-end
-
-local dispatchers = {
-  pr = dispatch_pr,
-  review = dispatch_review,
-  issue = dispatch_issue,
-  ci = dispatch_ci,
-  release = dispatch_release,
-  browse = dispatch_browse,
-  clear = function()
-    state_mod.clear_cache()
-    require('forge.logger').info('cache cleared')
-  end,
-}
 
 ---@param opts? forge.SurfaceNamesOpts
 ---@return string[]
@@ -1057,13 +628,7 @@ function M.parse(args, opts)
 end
 
 function M.dispatch(command)
-  local dispatcher = dispatchers[command.family]
-  if not dispatcher then
-    warn('unknown command: ' .. command.family)
-    return false
-  end
-  dispatcher(command)
-  return true
+  return require('forge.cmd_dispatch').dispatch(command)
 end
 
 function M.run(opts)
